requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
et-xmlfile==1.1.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.26.4
openai==1.30.3
openpyxl==3.1.2
pandas==2.2.2
playwright==1.44.0
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.2
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
